+++ src/packages/three/Three3Dtiles/Type.ts
@@ -1,0 +1,7 @@
+export interface Vec {
+  x: number
+  y: number
+  z: number
+}
+
+export type LngLat = [number, number]