+++ src/packages/three/Three3Dtiles/index.ts
@@ -1,0 +1,5 @@
+import { withInstall } from '../../../utils/with-install'
+import ThreeTiles3d from './ThreeTiles3d.vue'
+
+export const ElAmapThreeTiles3d = withInstall(ThreeTiles3d)
+export default ElAmapThreeTiles3d